[RESHAPE] from int32_t to uint8_t, from matrix-matrix, to matrix-vector. single tile working

+++ src/main/cpp/app/BitSerialMatMulAccelDriver.hpp
@@ -124,7 +124,7 @@ typedef struct {
 } HardwareCfg;
 
 typedef uint64_t PackedBitGroupType;
-typedef int32_t ResultType;
+typedef uint8_t ResultType;
 
 class BitSerialMatMulAccelDriver {
 public:
